formatted `Zig.py`, refactored inline errors, and fixed error regex
on zig 0.8.1, `zig fmt` does not output a leading './' if the file path
it was given did not include one, but the regex did not account for
this. furthermore, trying to close one inline error message would close
all error messages in all windows. now only the error message clicked on
is closed; however, there currently is no way to close all the inline error
messages in a single view. lastly we need to figure out the oldest version
of sublime text this package will support. backwards compatibility between
4 and 3 is quite good from what i can tell, but i'm not sure about version 2.